Add mc map
This maps event type names to production names.

Push it up as a json in format:

typename : [list of names of production macros]

for example there are two runs of the 2v - one simulating just the tail and the other using the whole lot, so we need to separate the event class from the way they were simulated in this file.